Removing perm
Perm was a valiant effort, and many good developers gave it their all. It was
time to amit that it was dead weight we were hauling along with us.

My main motivation was to get the permission querying code to be
something our vim setup could navigate through when going to the
definitions of functions named the same for db permissions and
perm. Once I got started I couldn't stop myself.

Assorted Notes:
- removed all traces of perm I could find. There are some instance of
"perm" in the codebase that have nothing to do with externalizing permissions.
- Made all perm config optional, as it's no longer used for anything.

Authored-by: Matthew Kocher <mkocher@pivotal.io>
