+++ app/actions/space_create.rb
@@ -3,8 +3,7 @@ module VCAP::CloudController
     class Error < ::StandardError
     end
 
-    def initialize(perm_client:, user_audit_info:)
-      @perm_client = perm_client
+    def initialize(user_audit_info:)
       @user_audit_info = user_audit_info
     end
 
@@ -16,10 +15,6 @@ module VCAP::CloudController
         Repositories::SpaceEventRepository.new.record_space_create(space, user_audit_info, message.audit_hash)
       end
 
-      VCAP::CloudController::Roles::SPACE_ROLE_NAMES.each do |role|
-        perm_client.create_space_role(role: role, space_id: space.guid)
-      end
-
       space
     rescue Sequel::ValidationFailed => e
       validation_error!(e)
@@ -27,7 +22,7 @@ module VCAP::CloudController
 
     private
 
-    attr_reader :perm_client, :user_audit_info
+    attr_reader :user_audit_info
 
     def validation_error!(error)
       if error.is_a?(Space::DBNameUniqueRaceError) || error.errors.on([:organization_id, :name])&.include?(:unique)
