+++ config/cloud_controller.yml
@@ -364,13 +364,6 @@ pending_builds:
   expiration_in_seconds: 42
   frequency_in_seconds: 300
 
-perm:
-  enabled: false
-  hostname: perm.service.cf.internal
-  port: 6283
-  ca_cert_path: "spec/fixtures/certs/perm_ca.crt"
-  timeout_in_milliseconds: 100
-
 credhub_api:
   internal_url: https://credhub.capi.internal:8844
   external_url: https://credhub.capi.external:8844
