+++ lib/cloud_controller/controller_factory.rb
@@ -25,7 +25,6 @@ module CloudController
     def default_dependencies
       {
         statsd_client: dependency_locator.statsd_client,
-        perm_client: dependency_locator.perm_client,
         object_renderer:     dependency_locator.object_renderer,
         collection_renderer: dependency_locator.paginated_collection_renderer,
       }
