--- lib/cloud_controller/perm/errors.rb
@@ -1,5 +1,0 @@
-module VCAP::CloudController
-  module Perm
-    class Error < StandardError; end
-  end
-end
