+++ lib/cloud_controller/permissions.rb
@@ -87,6 +87,23 @@ class VCAP::CloudController::Permissions
     roles.admin?
   end
 
+  def can_write_global_service_broker?
+    can_write_globally?
+  end
+
+  def can_write_space_scoped_service_broker?(space_guid)
+    can_write_to_space?(space_guid)
+  end
+
+  def can_read_space_scoped_service_broker?(service_broker)
+    service_broker.space_scoped? &&
+        can_read_secrets_in_space?(service_broker.space_guid, service_broker.space.organization_guid)
+  end
+
+  def can_read_service_broker?(service_broker)
+    can_read_globally? || can_read_space_scoped_service_broker?(service_broker)
+  end
+
   def readable_org_guids
     if can_read_globally?
       VCAP::CloudController::Organization.select(:guid).all.map(&:guid)
