+++ spec/spec_helper.rb
@@ -92,7 +92,7 @@ each_run_block = proc do
     rspec_config.mock_with :rspec do |mocks|
       mocks.verify_partial_doubles = true
     end
-    rspec_config.filter_run_excluding :opi, :perm, :stepper
+    rspec_config.filter_run_excluding :opi, :stepper
     rspec_config.expose_dsl_globally = false
     rspec_config.backtrace_exclusion_patterns = [%r{/gems/}, %r{/bin/rspec}]
