+++ spec/unit/access/organization_access_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 module VCAP::CloudController
   RSpec.describe OrganizationAccess, type: :access do
-    let(:queryer) { instance_spy(Permissions::Queryer) }
+    let(:queryer) { instance_spy(VCAP::CloudController::Permissions) }
 
     subject(:access) { OrganizationAccess.new(Security::AccessContext.new(queryer)) }
     let(:user) { VCAP::CloudController::User.make }
