+++ spec/unit/controllers/base/model_controller_spec.rb
@@ -8,7 +8,6 @@ module VCAP::CloudController
       {
         object_renderer: nil,
         collection_renderer: nil,
-        perm_client: double(Perm::Client),
         statsd_client: double(Statsd),
       }
     end
