+++ spec/unit/controllers/internal/log_access_controller_spec.rb
@@ -21,7 +21,6 @@ module VCAP::CloudController
         nil,
         {
           statsd_client: double(Statsd),
-          perm_client: double(Perm::Client)
         },
       )
     end
