+++ spec/unit/controllers/services/service_keys_controller_spec.rb
@@ -54,7 +54,6 @@ module VCAP::CloudController
           object_renderer: object_renderer,
           collection_renderer: collection_renderer,
           statsd_client: double(Statsd),
-          perm_client: double(Perm::Client)
         }
       end
       let(:config) { double(Config, get: nil) }
